Create 2025-09-04-AccountStatement.csv
forgot to upload the statement last night. markets are gapping up and as a result my stocks are also opening up 5bps. this will cause a bit of an issue with performance calculation, just like the vacation time off. At the end of the day I should not worry about this and look at the quarterly and yearly consistent performance.